+++ include/diy/algorithms.hpp
@@ -87,7 +87,7 @@ namespace diy
 
         typedef     diy::RegularContinuousLink      RCLink;
 
-        for (int i = 0; i < master.size(); ++i)
+        for (size_t i = 0; i < master.size(); ++i)
         {
             RCLink* link   = static_cast<RCLink*>(master.link(i));
             *link = RCLink(dim, domain, domain);
@@ -123,7 +123,7 @@ namespace diy
 
         // update master.expected to match the links
         int expected = 0;
-        for (int i = 0; i < master.size(); ++i)
+        for (size_t i = 0; i < master.size(); ++i)
             expected += master.link(i)->size_unique();
         master.set_expected(expected);
     }
@@ -150,7 +150,7 @@ namespace diy
 
         typedef     diy::RegularContinuousLink      RCLink;
 
-        for (int i = 0; i < master.size(); ++i)
+        for (size_t i = 0; i < master.size(); ++i)
         {
             RCLink* link   = static_cast<RCLink*>(master.link(i));
             *link = RCLink(dim, domain, domain);
@@ -186,7 +186,7 @@ namespace diy
 
         // update master.expected to match the links
         int expected = 0;
-        for (int i = 0; i < master.size(); ++i)
+        for (size_t i = 0; i < master.size(); ++i)
             expected += master.link(i)->size_unique();
         master.set_expected(expected);
     }
